move nuget stuff to a nuget package
